Add an external dependency to the "eggs" workspace
To test if Cachito will properly replace its version in the package.json
of the workspace.

The version must be replaced with the "version in Nexus", otherwise yarn
will not match the dependency to the resolved one in yarn.lock and will
try to fetch it from the original url.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>